requirements.txt
streamlit
beautifulsoup4
requests
email-validator
dnspython
pandas
